+++ ci/gpu/build.sh
@@ -45,10 +45,10 @@ gpuci_conda_retry install -y -c pytorch -c gwerbin \
     "cugraph=${MINOR_VERSION}" \
     "cuml=${MINOR_VERSION}" \
     "dask-cuda=${MINOR_VERSION}" \
-    "pytorch>=1.7" \
+    "pytorch=1.7.0" \
     "torchvision" \
     "python-confluent-kafka" \
-    "transformers" \
+    "transformers=3.5.*" \
     "seqeval=0.0.12" \
     "python-whois" \
     "requests" \